ADIÇÃO DE MAP NOS TYPES
